orness/requirements.txt
jsonschema>=4.23.0
python-dotenv>=1.0.1 
pandas
openpyxl==3.1.5
